% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bin_env}
\alias{bin_env}
\title{Helper function to prepare bin tables}
\usage{
bin_env(overall_range, M_range, sp_range, bin_size)
}
\arguments{
\item{overall_range}{(numeric) minimum and maximum values of all species and
Ms to be analyzed.}

\item{M_range}{matrix of ranges of environmental values in M for all species.
Columns must be minimum and maximum, and rows correspond to species.}

\item{sp_range}{matrix of ranges of environmental values in occurrences
for all species. Columns must be minimum and maximum, and rows correspond to
species.}

\item{bin_size}{(numeric) size of bins. Interval of values to be considered
when creating bin tables. Default = 10.}
}
\value{
A character matrix containing bins that identify accessibility of environments
(environemntal values inside M; "0"), used environments (presence of species
records; "1"), and unknown (uncertainty about the ability of the species to be
in such conditions given accessibility; "?").
}
\description{
Helper function to prepare bin tables
}
