% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ml_rec.R
\name{bin_ml_rec}
\alias{bin_ml_rec}
\title{Maximum likelihood reconstruction of characters}
\usage{
bin_ml_rec(tree_data)
}
\arguments{
\item{tree_data}{a list of two elements (phy and data) resulted from using the
function \code{\link[geiger]{treedata}}.}
}
\description{
Maximum likelihood reconstruction of characters
}
