% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_par_rec.R
\name{bin_par_rec}
\alias{bin_par_rec}
\title{Parsimony reconstruction of characters}
\usage{
bin_par_rec(tree_data)
}
\arguments{
\item{tree_data}{a list of two elements (phy and data) resulted from using the
function \code{\link[geiger]{treedata}}.}
}
\description{
Parsimony reconstruction of characters
}
