% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_tables0.R
\name{bin_tables0}
\alias{bin_tables0}
\title{Bin tables of environmental conditions in M and occurrences}
\usage{
bin_tables0(M_folder, M_format, occ_folder, longitude, latitude,
  var_folder, var_format, round = FALSE, round_names,
  multiplication_factor = 1, percentage_out = 5, bin_size = 10,
  output_directory = "Species_E_bins")
}
\arguments{
\item{M_folder}{(character) name of the folder containing files representing
the accessible area (M) for all species to be analyzed.}

\item{M_format}{format of files representing the accessible area (M) for the
species. Names must match with the ones for occurrence files in
\code{occ_folder}. Options are: "shp", "gpkg", or any of the options in
\code{\link[raster]{writeFormats}}.}

\item{occ_folder}{(character) name of the folder containing csv files of
occurrence data for all species.}

\item{longitude}{(character) name of the column containing values of longitude
of occurrences.}

\item{latitude}{(character) name of the column containing values of latitude
of occurrences.}

\item{var_folder}{(character) name of the folder containing layers to
represent environmental variables.}

\item{var_format}{format of layers to represent environmental variables. See
options in \code{\link[raster]{writeFormats}}.}

\item{round}{(logical) whether or not to round the values of one or more
variables after multiplying them times \code{multiplication_factor}.
Default = FALSE.}

\item{round_names}{(character) names of the variables to be rounded.
Default = NULL. If \code{round} = TRUE, names must be defined.}

\item{multiplication_factor}{(numeric) value to be used to multiply the
variables defined in \code{round_names}. Default = 1.}

\item{percentage_out}{(numeric) percentage of extreme environmental data in M
to be excluded in bin creation for further analyses. Default = 5.}

\item{bin_size}{(numeric) size of bins. Interval of values to be considered
when creating bin tables. Default = 10.}

\item{output_directory}{(character) name of the folder in which results will be
written. Default = "Species_E_bins".}
}
\value{
A folder named as in \code{output_directory} containing all resultant csv
files for all variables, with bins for all species. Results will also be
returned as a list.
}
\description{
bin_tables0 helps in creating csv files with bin tables
of environmental conditions in M and species occurrence records. All of this
starting from raw data, for various species, and using multiple variables.
}
