% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_tables_null.R
\name{bin_tables_null}
\alias{bin_tables_null}
\title{Bin tables of environemntal conditions for virtual species}
\usage{
bin_tables_null(M_folder, M_format, var_folder, var_format,
  round = FALSE, round_names = NULL, multiplication_factor = 1,
  bin_size = 10, output_directory = "Virtual_species_E_bins")
}
\arguments{
\item{M_folder}{(character) name of the folder containing files representing
the accessible area (M) for all species to be analyzed.}

\item{M_format}{format of files representing the accessible area (M) for the
species. Options are: "shp", "gpkg", or any of the options in
\code{\link[raster]{writeFormats}}.}

\item{var_folder}{(character) name of the folder conatining layers to
represent environmental variables.}

\item{var_format}{format of layers to represent environmental variables. See
options in \code{\link[raster]{writeFormats}}.}

\item{round}{(logical) whether or not to round the values of one or more
variables after multiplying them times \code{multiplication_factor}.
Default = FALSE.}

\item{round_names}{(character) names of the variables to be rounded.
Default = NULL. If \code{round} = TRUE, names must be defined.}

\item{multiplication_factor}{(numeric) value to be used to multiply the
variables defined in \code{round_names}. Default = 1.}

\item{bin_size}{(numeric) size of bins. Interval of values to be considered
when creating bin tables. Default = 10.}

\item{output_directory}{(character) name of the folder in which results will be
written. Default = "Virtual_species_E_bins".}
}
\value{
A folder named as in \code{output_directory} containing all resultant csv
files for all variables, with bins for all species. Results will also be
returned as a list.
}
\description{
bin_tables_null helps in creating csv files with bin tables
of environmental conditions for virtual species. All of this for various
species and multiple variables.
}
