% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{character_table}
\alias{character_table}
\title{Example of character table for six species}
\format{A character matrix with 6 rows and 28 columns.}
\usage{
character_table
}
\description{
A character table representing species ecological niches derived from
previous preparation processes. Each row represents a species and each column
a binary character in which one or more values of the environmental variable
are categorized as used "1", non used "0", or uncertain "?".
}
\examples{
data("character_table", package = "nichevol")

head(character_table)
}
\keyword{datasets}
