% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms_env.R
\name{histograms_env}
\alias{histograms_env}
\title{Histograms of environmental conditions at occurrences and in M}
\usage{
histograms_env(M_folder, M_format, occ_folder, longitude, latitude,
  var_folder, var_format, CL_lines = c(95, 99), col = NULL,
  round = FALSE, round_names = NULL, multiplication_factor = 1,
  save_ranges = TRUE, output_directory = "Histogram_ranges")
}
\arguments{
\item{M_folder}{(character) name of the folder containing files representing
the accessible area (M) for all species to be analyzed.}

\item{M_format}{format of files representing the accessible area (M) for the
species. Names must match with the ones for occurrence files in
\code{occ_folder}. Options are: "shp", "gpkg", or any of the options in
\code{\link[raster]{writeFormats}}.}

\item{occ_folder}{(character) name of the folder containing csv files of
occurrence data for all species.}

\item{longitude}{(character) name of the column containing values of longitude
of occurrences.}

\item{latitude}{(character) name of the column containing values of latitude
of occurrences.}

\item{var_folder}{(character) name of the folder containing layers to
represent environmental variables.}

\item{var_format}{format of layers to represent environmental variables. See
options in \code{\link[raster]{writeFormats}}.}

\item{CL_lines}{(numeric) confidence limits to be plotted in the histograms.
Default = c(95, 99).}

\item{col}{colors for lines representing confidence limits. If NULL, colors
are selected from a gray palette. Default = NULL.}

\item{round}{(logical) whether or not to round the values of one or more
variables after multiplying them times \code{multiplication_factor}.
Default = FALSE.}

\item{round_names}{(character) names of the variables to be rounded.
Default = NULL. If \code{round} = TRUE, names must be defined.}

\item{multiplication_factor}{(numeric) value to be used to multiply the
variables defined in \code{round_names}. Default = 1.}

\item{save_ranges}{(logical) whether or not to save the values identified as
ranges considering the whole set of values and confidence limits defined in
\code{CL_lines}. Default = TRUE.}

\item{output_directory}{(character) name of the folder in which results will be
written. Default = "Histogram_ranges".}
}
\value{
A folder named as in \code{output_directory} containing all resultant PDF
files for all variables, with histograms for all species. Results will also be
returned as a list.
}
\description{
histograms_env creates PDF files with histogram plots of
environmental conditions in M, lines for the confidence limits of values in
M, and the location of values in occurrence records for multiple species and
multiple variables.
}
