% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{m}
\alias{m}
\alias{m1}
\alias{m2}
\alias{m3}
\alias{m4}
\alias{m5}
\alias{m6}
\title{Example of accessible areas for a species}
\format{
A SpatVector object.
}
\value{
No return value, used with function \code{\link[terra]{vect}} to
bring an example of an accessible area for a species.
}
\description{
A SpatVector object representing the accessible area for a species.
}
\examples{
m1 <- terra::vect(system.file("extdata", "m1.gpkg", package = "nichevol"))

terra::plot(m1)
}
