% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{m_list}
\alias{m_list}
\title{Example of accessible areas for six species}
\format{A list of 6 SpatialPolygonsDataFrame objects.}
\usage{
m_list
}
\description{
A list of 6 elements (SpatialPolygonsDataFrame objects) representing
accessible areas for 6 species.
}
\examples{
data("m_list", package = "nichevol")

str(m_list)
}
\keyword{datasets}
