% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_bars.R
\name{niche_bars}
\alias{niche_bars}
\title{Bars for niche shifts among distinct taxa}
\usage{
niche_bars(bin_table, present = "1", unknown = "?",
  present_col = "red4", unknown_col = "grey85",
  absent_col = "royalblue1", width = 50, height = 5, res = 300,
  output_directory = "Niche_bars")
}
\arguments{
\item{bin_table}{matrix of environmental bins for all species derived from
using the functions \code{\link{bin_tables}} or \code{\link{bin_tables0}}.}

\item{present}{(character) code indicating environmental bins in which the
species is present. Default = "1".}

\item{unknown}{(character) code indicating environmental bins in which the
species presence is unknown (uncertain). Default = "?".}

\item{present_col}{color for line representing environments where the species
is present. Default = "red4".}

\item{unknown_col}{color for line representing environments where the species
presence is unknown (uncertain). Default = "grey85".}

\item{absent_col}{color for area of the bar representing environments where
no change has been detected. Default = "royalblue1".}

\item{width}{(numeric) width of the device in mm to be passed to the
\code{\link[grDevices]{png}} function. Default = 50.}

\item{height}{(numeric) height of the device in mm to be passed to the
\code{\link[grDevices]{png}} function. Default = 5.}

\item{res}{(numeric) nominal resolution in ppi to be passed to the
\code{\link[grDevices]{png}} function. Default = 300.}

\item{output_directory}{(character) name of the folder in which results will
be written. The directory will be created as part of the process.
Default = "Niche_bars"}
}
\value{
A folder named as in \code{output_directory} containing all bar figures
produced.
}
\description{
niche_bars produces bar figures that represent how
species niches have changed compared to others.
}
