% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{niche_labels}
\alias{niche_labels}
\title{Labels to represent niches of tips and ancestors}
\usage{
niche_labels(tree, whole_rec_table, label_type = "tip_node",
  tip_offset = 0.015, present = "1", unknown = "?",
  present_col = "#e41a1c", unknown_col = "#969696",
  absent_col = "#984ea3", width = 1, height = 1)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{whole_rec_table}{matrix of environmental bins for all tips and nodes
derived from functions \code{\link{bin_par_rec}} or \code{\link{bin_ml_rec}}.}

\item{label_type}{(character) type of label; options are: "tip", "node", and
"tip_node". Default = "tip_node".}

\item{tip_offset}{(numeric) space between tips and the labels. Default = 0.015.}

\item{present}{(character) code indicating environmental bins in which the
species is present. Default = "1".}

\item{unknown}{(character) code indicating environmental bins in which the
species presence is unknown (uncertain). Default = "?".}

\item{present_col}{color for area of the bar representing environments where
the species is present. Default = "#e41a1c".}

\item{unknown_col}{color for area of the bar representing environments where
the species presence is unknown (uncertain). Default = "#969696".}

\item{absent_col}{color for area of the bar representing environments where
no change has been detected. Default = "#984ea3".}

\item{width}{value defining the width of niche bars; default = 1.}

\item{height}{value defining the height of niche bars; default = 1.}
}
\description{
niche_labels helps in adding bar-type labels that represent how
species niches are compared to others.
}
