% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{niche_legend}
\alias{niche_legend}
\title{Legends for niche labels in phylogenetic trees}
\usage{
niche_legend(position, legend = c("Uncertain", "Present", "Not present"),
  pch = 22, pt.bg = c("#969696", "#e41a1c", "#984ea3"),
  col = "transparent", pt.cex = 2.2, bty = "n", ...)
}
\arguments{
\item{position}{(character or numeric) position of legend. If character,
part of the plot (e.g., "topleft"), see \code{\link[graphics]{legend}}. If
numeric, vector of two values indicating x and y postion (e.g., c(0.1, 6)).}

\item{legend}{(character) vector of length = three indicating the text to
identify environments with uncertain presence, presence, and absence of the
species. Default = c("Uncertain", "Present", "Not present").}

\item{pch}{point type as in \code{\link[graphics]{points}}. Default = 22.}

\item{pt.bg}{colors to represent what is in \code{legend}.
Default = c("#969696", "#e41a1c", "#984ea3").}

\item{col}{border of symbol (points). Default = "transparent".}

\item{pt.cex}{size of symbol (points). Default = 2.2.}

\item{bty}{legend border type. Default = "n".}

\item{...}{Other arguments from function \code{\link[graphics]{legend}} other
than the ones described above.}
}
\description{
Legends for niche labels in phylogenetic trees
}
