% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichevol.R
\docType{package}
\name{nichevol}
\alias{nichevol}
\alias{nichevol-package}
\title{nichevol: An R package for assessment of evolutionary change of ecological niches}
\description{
nichevol implements multiple tools to help in performing critical steps of the
process of assessment of evolution of species ecological niches. The main
analyses include: intial exploration of environmental data in species records
and accessible areas, preparation of data for phylogenetic analyses,
phylogenetical analyses, and plotting for interpretations.
}
\section{Main functions in nichevol}{

\code{\link{bin_tables}}, \code{\link{bin_tables0}},
\code{\link{bin_tables_null}}, \code{\link{histograms_env}},
\code{\link{niche_bars}}, \code{\link{nichevol_bars}}

Other functions (important helpers)

\code{\link{bin_env}}, \code{\link{bin_env_null}},
\code{\link{pdf_histograms}}, \code{\link{rformat_type}}
}

