% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_diffbars.R
\name{nichevol_bars}
\alias{nichevol_bars}
\title{Bars for niche shifts among distinct taxa}
\usage{
nichevol_bars(reconstructed_bins, species_rows, present = "1",
  absent = "0", unknown = "0 1", present_col = "grey10",
  unknown_col = "white", no_change_col = "grey90",
  retraction_col = "dodgerblue3", expansion_col = "green1",
  width = 50, height = 5, res = 300,
  output_directory = "Difference_bars")
}
\arguments{
\item{reconstructed_bins}{matrix of reconstructed bins for nodes and species
derived from a process of maximum parsimony reconstruction.}

\item{species_rows}{(numeric) vector indicating the rows of the matrix in
which species bins are.}

\item{present}{(character) code indicating environmental bins in which the
species is present. Default = "1".}

\item{absent}{(character) code indicating environmental bins in which the
species is absent. Default = "0".}

\item{unknown}{(character) code indicating environmental bins in which the
species presence is unknown (uncertain). Default = "0 1".}

\item{present_col}{color for line representing environments where the species
is present. Default = "grey10".}

\item{unknown_col}{color for line representing environments where the species
presence is unknown (uncertain). Default = "white".}

\item{no_change_col}{color for area of the bar representing environments where
no change has been detected. Default = "grey90".}

\item{retraction_col}{color for area of the bar representing environments where
niche retraction has been detected. Default = "dodgerblue3".}

\item{expansion_col}{color for area of the bar representing environments where
niche expansion has been detected. Default = "green1".}

\item{width}{(numeric) width of the device in mm to be passed to the
\code{\link[grDevices]{png}} function. Default = 50.}

\item{height}{(numeric) height of the device in mm to be passed to the
\code{\link[grDevices]{png}} function. Default = 5.}

\item{res}{(numeric) nominal resolution in ppi to be passed to the
\code{\link[grDevices]{png}} function. Default = 300.}

\item{output_directory}{(character) name of the folder in which results will
be written. The directory will be created as part of the process.
Default = "Difference_bars"}
}
\value{
A folder named as in \code{output_directory} containing all bar figures
produced.
}
\description{
nichevol_bars helps in producing bar figures that represent how
species niches have changed compared to others.
}
