% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{nichevol_labels}
\alias{nichevol_labels}
\title{Labels to represent changes of between ancestors and descendants}
\usage{
nichevol_labels(tree, whole_rec_table, ancestor_line = FALSE,
  present = "1", absent = "0", unknown = "?",
  present_col = "#252525", unknown_col = "#d9d9d9",
  no_change_col = "#b2df8a", retraction_col = "#984ea3",
  expansion_col = "#4daf4a", width = 1, height = 1)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{whole_rec_table}{matrix of reconstructed bins for nodes and species
derived from a process of maximum parsimony reconstruction.}

\item{ancestor_line}{controls whether ancestor line is plotted.
Default = FALSE.}

\item{present}{(character) code indicating environmental bins in which the
species is present. Default = "1".}

\item{absent}{(character) code indicating environmental bins in which the
species is absent. Default = "0".}

\item{unknown}{(character) code indicating environmental bins in which the
species presence is unknown (uncertain). Default = "?".}

\item{present_col}{color for line representing environments where the species
is present. Default = "#252525".}

\item{unknown_col}{color for line representing environments where the species
presence is unknown (uncertain). Default = "#d9d9d9".}

\item{no_change_col}{color for area of the bar representing environments where
no change has been detected. Default = "#b2df8a".}

\item{retraction_col}{color for area of the bar representing environments where
niche retraction has been detected. Default = "#984ea3".}

\item{expansion_col}{color for area of the bar representing environments where
niche expansion has been detected. Default = "#4daf4a".}

\item{width}{value defining the width of bars representing changes in niches;
default = 1.}

\item{height}{value defining the height of bars representing changes in niches;
default = 1.}
}
\description{
nichevol_labels helps in adding bar-type labels that represent how
species niches changed from ancestors to descendants.
}
