% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{nichevol_legend}
\alias{nichevol_legend}
\title{Legends for niche evolution labels in phylogenetic trees}
\usage{
nichevol_legend(position, legend = c("Uncertain", "Present", "No change",
  "Retraction", "Expansion"), pch = 22, pt.cex = 2.2, lty = 1,
  lwd = 1, col = c("orange", "grey10", "grey90", "dodgerblue3",
  "green1"), cex = 1, bty = "n", ...)
}
\arguments{
\item{position}{(character or numeric) position of legend. If character,
part of the plot (e.g., "topleft"), see \code{\link[graphics]{legend}}. If
numeric, vector of two values indicating x and y postion (e.g., c(0.1, 6)).}

\item{legend}{(character) vector of length = five indicating the text to
identify environments with uncertain presence and presence of the species,
as well as, areas where niches have not changed, have retracted or expanded.
Default = c("Uncertain", "Present", "No change", "Retraction", "Expansion").
Order must be mantained.}

\item{pch}{point type as in \code{\link[graphics]{points}}. Default = 22.}

\item{pt.cex}{size of symbol (points). Default = 2.2.}

\item{lty}{line type see \code{\link[graphics]{par}}. Default = 1.}

\item{lwd}{line width see \code{\link[graphics]{par}}. Default = 1.}

\item{col}{vector of five colors to represent what is in legend.
Default = c("orange", "grey10", "grey90", "dodgerblue3", "green1").}

\item{cex}{size of all elements in legend see \code{\link[graphics]{par}}.
Deafult = 1.}

\item{bty}{legend border type. Default = "n".}

\item{...}{Other arguments from function \code{\link[graphics]{legend}} other
than the ones described above.}
}
\description{
Legends for niche evolution labels in phylogenetic trees
}
