% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{occ_list}
\alias{occ_list}
\title{Example of occurrence records for six species}
\format{A list of 6 data.frames:
\describe{
\item{species}{species name, a code in this example}
\item{x}{longitude, longitude value}
\item{y}{latitude, latitude value}
}}
\usage{
occ_list
}
\description{
A list of 6 data.frames containing name and geographic coordinates for
6 species.
}
\examples{
data("occ_list", package = "nichevol")

str(occ_list)
}
\keyword{datasets}
