% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pdf_histograms}
\alias{pdf_histograms}
\title{Helper function to create PDF files with histograms}
\usage{
pdf_histograms(env_data, occ_data, y_values, sp_names, variable_name,
  CL_lines, limits, col, output_directory = "Histogram_ranges_check")
}
\arguments{
\item{env_data}{list of environmental values in M for all species.}

\item{occ_data}{list of environmental values in occurrences for all species.}

\item{y_values}{list of values for the y axis to be used to represent where
occurrences are distributed across the environmental values in M.}

\item{sp_names}{(character) names of the species for which the process will
be performed.}

\item{variable_name}{(character) name of the variable to be plotted.}

\item{CL_lines}{(numeric) confidence limits to be plotted in the histograms.}

\item{limits}{numeric matrix containing the actual values for the confidence
limits of M.}

\item{col}{color for lines representing the confidence limits of M.}

\item{output_directory}{(character) name of the folder in which results will be
written. Default = "Histogram_ranges_check".}
}
\value{
A PDF file written in the output directory containing all resultant figures.
}
\description{
Helper function to create PDF files with histograms
}
