% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_polygons.R
\name{random_polygons}
\alias{random_polygons}
\title{Generation of random polygons in a given area}
\usage{
random_polygons(polygon, style = "TR", n_polygons = 100,
  n_vertices = 25, minimum_distance = 10, length_threshold = 5,
  buffer_distance = 0, overwrite = FALSE,
  output_directory = "Random_polygons")
}
\arguments{
\item{polygon}{SpatialPolygonsDataFrame object. CRS WGS84 is required.}

\item{style}{(character) algorithm to be used when creating polygons. Options
are: "TR" for vertices randomly located across the entire area; and "BR" for
vertices placed randomly across the entire area and in nine blocks derived
from dividing the area in equal number of random points. Default = "TR".}

\item{n_polygons}{(numeric) number of polygons to be created; default = 100.}

\item{n_vertices}{(numeric) maximum number of vertices for polygons.}

\item{minimum_distance}{(numeric) approximate minimum distance in km for
separation among vertices. Default = 10.}

\item{length_threshold}{(numeric) approximate distance in km for producing
concavity in polygons. Default = 5.}

\item{buffer_distance}{(numeric) approximate distance in km to buffer
resultant polygons. Default = 0.}

\item{overwrite}{(logical) whether or not to overwrite previous results.}

\item{output_directory}{(character) name of the folder in which results will
be written. Default = "Random_polygons".}
}
\value{
A folder named as in \code{output_directory} containing all resultant
shapefiles of random polygons. Results will also be returned as a list of
spatial objects.
}
\description{
random_polygons creates polygons of random size and complexity within
a given SpatialPolygonsDataFrame, trying to fill the area with the resultant polygons
in at least nine quadrants. This is designed to simulate virtual species' Ms (aka
"training" or "background" regions).
}
\details{
Distances are approximate because 1 decimal degree is assumed to equal 111.32
km.

Style for random polygons "BR" may help to get smaller and more uniformly
distributed across the area.
}
