% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_table.R
\name{read_bin_table}
\alias{read_bin_table}
\alias{read_bin_tables}
\title{Read tables of binary niche characters from directory}
\usage{
read_bin_table(file)

read_bin_tables(directory)
}
\arguments{
\item{file}{(character) name of CSV file containing a table of binary niche
characters.}

\item{directory}{(character) name of directory where tables of binary niche
characters were written as CSV files.}
}
\value{
A matrix if \code{read_bin_table} is used.

A list of matrices if \code{read_bin_tables} is used.
}
\description{
Read one or multiple tables binary niche characters from
directory.
}
