% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rename_tips}
\alias{rename_tips}
\title{Helper function to rename tips of trees}
\usage{
rename_tips(tree, names)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{names}{(character) vector of new names. Length must be equal to number
of tips.}
}
\description{
Helper function to rename tips of trees
}
