% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rformat_type}
\alias{rformat_type}
\title{Helper function to find raster extention}
\usage{
rformat_type(format)
}
\arguments{
\item{format}{(character) any of the format types allowed for raster objects.
See \code{\link[raster]{writeFormats}}}
}
\value{
Raster extension according to format type.
}
\description{
Helper function to find raster extention
}
