% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{score_tree}
\alias{score_tree}
\title{Helper function to assign bin scores to every tip in a given tree}
\usage{
score_tree(tree, character_table, species_col, unknown = FALSE)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{character_table}{data.frame containig characters for all species.}

\item{species_col}{(character) name of the column containing name of species.}

\item{unknown}{(logical) whether or not there are unknown tips.}
}
\description{
Helper function to assign bin scores to every tip in a given tree
}
