% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sig_sq}
\alias{sig_sq}
\title{Helper function to get sigma squared values for a given dataset}
\usage{
sig_sq(tree_data, model = "BM")
}
\arguments{
\item{tree_data}{a list of two elements (phy and data) resulted from using the
function \code{\link[geiger]{treedata}}.}

\item{model}{model to fit to comparative data; see
\code{\link[geiger]{fitContinuous}}. Default = "BM".}
}
\description{
Helper function to get sigma squared values for a given dataset
}
