% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_rec.R
\name{smooth_rec}
\alias{smooth_rec}
\title{Parsimony reconstruction of characters}
\usage{
smooth_rec(whole_rec_table)
}
\arguments{
\item{whole_rec_table}{matrix containing all reconstructed charaters for all
tips and nodes. It results from using the functions \code{\link{bin_par_rec}}
or \code{\link{bin_ml_rec}}.}
}
\description{
Parsimony reconstruction of characters
}
