% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_evalues_virtual.R
\name{stats_evalues_virtual}
\alias{stats_evalues_virtual}
\title{Statistics of environmental conditions for virtual species}
\usage{
stats_evalues_virtual(stats = c("median", "range"), virtualsp_folder,
  virtualsp_format, var_folder, var_format, round = FALSE, round_names,
  multiplication_factor = 1, percentage_out = 0,
  output_directory = "Virtual__species_E_stats")
}
\arguments{
\item{stats}{(character) name or vector of names of functions to be applied
to get basic statistics of environmental values. Options include: "mean", "sd",
"var", "median", "range", "quantile".}

\item{virtualsp_folder}{(character) name of the folder containing files
representing the virtually occupied area of all species to be analyzed.}

\item{virtualsp_format}{format of files representing the virtually occupied
area of all species. Options are: "shp", "gpkg", or any of the options in
\code{\link[raster]{writeFormats}}.}

\item{var_folder}{(character) name of the folder containing layers to
represent environmental variables.}

\item{var_format}{format of layers to represent environmental variables. See
options in \code{\link[raster]{writeFormats}}.}

\item{round}{(logical) whether or not to round the values of one or more
variables after multiplying them times \code{multiplication_factor}.
Default = FALSE.}

\item{round_names}{(character) names of the variables to be rounded.
Default = NULL. If \code{round} = TRUE, names must be defined.}

\item{multiplication_factor}{(numeric) value to be used to multiply the
variables defined in \code{round_names}. Default = 1.}

\item{percentage_out}{(numeric) percentage of extreme environmental data in M
to be excluded in bin creation for further analyses. Default = 5.}

\item{output_directory}{(character) name of the folder in which results will
be written. Default = "Species_E_stats".}
}
\value{
A folder named as in \code{output_directory} containing all resultant csv
files for all variables, with statistics for all species. Results will also be
returned as a list.
}
\description{
stats_evalues_virtual creates csv files with statistics
of environmental conditions for multiple virtual species and multiple
variables.
}
