% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{temp}
\alias{temp}
\title{Example of an environmental variable used in analysis}
\format{
A SpatRaster object.
}
\value{
No return value, used with function \code{\link[terra]{rast}} to
bring an example of an environmental variable used in analysis.
}
\description{
A SpatRaster object representing the variable temperature.
}
\examples{
temp <- terra::rast(system.file("extdata", "temp.tif", package = "nichevol"))

terra::plot(temp)
}
