% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tree}
\alias{tree}
\title{Example of a phylogenetic tree for six species}
\format{An object of class phylo for 6 species.}
\usage{
tree
}
\description{
A phylogenetic tree with 6 species and their relationships.
}
\examples{
data("tree", package = "nichevol")

str(tree)
}
\keyword{datasets}
