% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tree5}
\alias{tree5}
\title{Example of a phylogenetic tree for five species}
\format{An object of class phylo for 5 species.}
\usage{
tree5
}
\description{
A phylogenetic tree with 5 species and their relationships.
}
\examples{
data("tree5", package = "nichevol")

str(tree5)
}
\keyword{datasets}
