% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tree_data}
\alias{tree_data}
\title{Example of a list containing a tree and a table of characters for six species}
\format{A list of 2 elements:
\describe{
\item{phy}{object of class phylo for 6 species}
\item{data}{matrix of 6 rows and 28 columns}
}}
\usage{
tree_data
}
\description{
A list of 2 elements (phy and data) resulting from using the function
\code{\link[geiger]{treedata}}.
}
\examples{
data("tree_data", package = "nichevol")

str(tree_data)
}
\keyword{datasets}
