% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bin_env_null}
\alias{bin_env_null}
\title{Helper function to prepare bin tables of null species}
\usage{
bin_env_null(overall_range, M_range, bin_size)
}
\arguments{
\item{overall_range}{(numeric) minimum and maximum values of all species and
Ms to be analyzed.}

\item{M_range}{matrix of environmental values ranges in M for all species.
Columns must be minimum and maximum, and rows correspond to species.}

\item{bin_size}{(numeric) size of bins. Interval of values to be considered
when creating bin tables. Default = 10.}
}
\value{
A character matrix containing bins that identify accessibility of environments
(environmental values inside M; "0"), used environments (presence of species
records; "1"), and unknown (uncertainty about the ability of the species to be
in such conditions given accessibility; "?").
}
\description{
Helper function to prepare bin tables of null species
}
