% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ml_rec.R
\name{bin_ml_rec}
\alias{bin_ml_rec}
\title{Maximum likelihood reconstruction of characters}
\usage{
bin_ml_rec(tree_data)
}
\arguments{
\item{tree_data}{a list of two elements (phy and data) resulting from using the
function \code{\link[geiger]{treedata}}.}
}
\value{
A table with columns representing bins, rows representing first tip states and then reconstructed nodes
}
\description{
Maximum likelihood reconstruction of characters
}
\examples{

tree <- phytools::pbtree(b = 1, d = 0, n = 5, scale = TRUE,
                         nsim = 1, type = "continuous", set.seed(5));
dataTable <- cbind("241" = rep("1", length(tree$tip.label)),
                   "242" = rep("1", length(tree$tip.label)),
                   "243" = c("1", "1", "0", "0", "0"),
                   "244" = c("1", "1", "0", "0", "0"),
                   "245" = c("1", "?", "0", "0", "0"));
rownames(dataTable) <- tree$tip.label;
treeWdata <- geiger::treedata(tree, dataTable);
bin_ml_rec(treeWdata);

}
