% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_tables.R
\name{bin_tables}
\alias{bin_tables}
\title{Bin tables of environemntal conditions in M and occurrences}
\usage{
bin_tables(ranges, percentage_out = 5, bin_size = 10,
  output_directory = "Species_E_bins")
}
\arguments{
\item{ranges}{list of ranges of environmental values in M and in species
occurrences derived from using the function \code{\link{histograms_env}}.}

\item{percentage_out}{(numeric) percentage of extreme environmental data in M
to be excluded in bin creation for further analyses. Default = 5.}

\item{bin_size}{(numeric) size of bins. Interval of values to be considered
when creating bin tables. Default = 10.}

\item{output_directory}{(character) name of the folder in which results will be
written. Default = "Species_E_bins".}
}
\value{
A folder named as in \code{output_directory} containing all resultant csv
files for all variables, with bins for all species. Results will also be
returned as a list.
}
\description{
bin_tables0 helps in creating csv files with bin tables
of environmental conditions in M and species occurrence records. All of this
starting from results of previous analyses, for various species, and using
multiple variables.
}
