% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_9blocks}
\alias{make_9blocks}
\title{Helper function to split geographic points in 9 blocks of equal size}
\usage{
make_9blocks(data)
}
\arguments{
\item{data}{matrix with longitude and latitude columns, in that order.}
}
\description{
Helper function to split geographic points in 9 blocks of equal size
}
\examples{
# random data
rdata <- cbind(x = rnorm(100, -80, 12), y = rnorm(100, -3, 15))
blocks <- make_9blocks(rdata)
}
