% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{niche_labels}
\alias{niche_labels}
\title{Labels to represent niches of tips and ancestors}
\usage{
niche_labels(whole_rec_table, label_type = c("tip", "node", "tip_node"),
  tip_offset = 0.015, present = "1", unknown = "?",
  present_col = "red4", unknown_col = "lightblue",
  absent_col = "royalblue1")
}
\arguments{
\item{whole_rec_table}{matrix of environmental bins for all tips and nodes
derived from functions \code{\link{bin_par_rec}} or \code{\link{bin_ml_rec}}.}

\item{present}{(character) code indicating environmental bins in which the
species is present. Default = "1".}

\item{unknown}{(character) code indicating environmental bins in which the
species presence is unknown (uncertain). Default = "?".}

\item{present_col}{color for area of the bar representing environments where
the species is present. Default = "red4".}

\item{unknown_col}{color for area of the bar representing environments where
the species presence is unknown (uncertain). Default = "lightblue".}

\item{absent_col}{color for area of the bar representing environments where
no change has been detected. Default = "royalblue1".}
}
\description{
niche_labels helps in adding bar-type labels that represent how
species niches are compared to others.
}
