% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{nichevol_labels}
\alias{nichevol_labels}
\title{Labels to represent changes of between ancestors and descendants}
\usage{
nichevol_labels(whole_rec_table, ancestor_line = FALSE, present = "1",
  absent = "0", unknown = "?", present_col = "grey10",
  unknown_col = "orange", no_change_col = "grey90",
  retraction_col = "dodgerblue3", expansion_col = "green1")
}
\arguments{
\item{whole_rec_table}{matrix of reconstructed bins for nodes and species
derived from a process of maximum parsimony reconstruction.}

\item{present}{(character) code indicating environmental bins in which the
species is present. Default = "1".}

\item{absent}{(character) code indicating environmental bins in which the
species is absent. Default = "0".}

\item{unknown}{(character) code indicating environmental bins in which the
species presence is unknown (uncertain). Default = "?".}

\item{present_col}{color for line representing environments where the species
is present. Default = "grey10".}

\item{unknown_col}{color for line representing environments where the species
presence is unknown (uncertain). Default = "white".}

\item{no_change_col}{color for area of the bar representing environments where
no change has been detected. Default = "grey90".}

\item{retraction_col}{color for area of the bar representing environments where
niche retraction has been detected. Default = "dodgerblue3".}

\item{expansion_col}{color for area of the bar representing environments where
niche expansion has been detected. Default = "green1".}
}
\description{
nichevol_labels helps in adding bar-type labels that represent how
species niches changed from ancestors to descendants.
}
