% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rename_tips}
\alias{rename_tips}
\title{Helper function to rename tips of trees for simulations}
\usage{
rename_tips(tree, names)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{names}{(character) vector of new names. Length must be equal to number
of tips. They will be assigned in the order given.}
}
\value{
Tree of class "phylo" with specified names
}
\description{
Helper function to rename tips of trees for simulations
}
\examples{
# a simple tree
data("tree5", package = "nichevol")

# renaming tips
renamedTree <- rename_tips(tree5, c("a", "b", "c", "d", "e"))
}
