% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{score_tip}
\alias{score_tip}
\title{Helper function to calculate the median bin score for a given species}
\usage{
score_tip(character_table, species_col, species_name, unknown = FALSE)
}
\arguments{
\item{character_table}{data.frame containig characters for all species.}

\item{species_col}{(character) name of the column containing name of species.}

\item{species_name}{(character) name of the species to be analyzed.}

\item{unknown}{(logical) whether or not there are unknown tips.}

\item{tree}{an object of class "phylo".}
}
\description{
Helper function to calculate the median bin score for a given species
}
